Inspector: Use the request url instead of the response url to find resources
The request url and the response url are identical unless there is a redirect.
However, when trying to find a resource in MemoryCache or ResourceFetcher, it
is only stored based on the request url, so a redirected resource may not be
found in the cache if searched based on the response url.

Update InspectorResourceContentLoader to use a stylesheet's href, which was
the requested url, instead of the base URL, which was the response url.

Update NetworkResourcesData to store the request url instead of the response
url, since it is only used to match against a request url.

BUG=

Review URL: https://codereview.chromium.org/1694883002

Cr-Original-Commit-Position: refs/heads/master@{#378344}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f364738f042e8b817450646b6a7e6f22ae954c95