[Line Layout API] Start converting LayoutBlockFlowLine uses of LayoutObject* operator.
LayoutBlockFlowLine is turning out to be a bear. This patch is mostly
mechanical but the following one will not be, so I split it up here.

This involved using the LineLayoutAPIShim in AXObjectCacheImpl.

Separately, a minor operator removal in LineLayoutSVGInlineText.h.

BUG=499321

Review URL: https://codereview.chromium.org/1731103004

Cr-Original-Commit-Position: refs/heads/master@{#378587}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 7a2979b2b929465d5b65920825ea72f7ae65a47b