Rework how --target, --debug, and --release work in run-webkit-tests.
This CL will enable run-webkit-tests to work correctly with
build directories that do not end in "Debug" or "Release"; this
most obviously affects GN builds (which can be in directories with
arbitrary names) but will also affect win x64 builds (which used
Debug_x64 and Release_x64).

Without this fix, any non-standard directory would work if you
specified it with --target (or --build-directory + --target), but
any relevant TestExpectations would be ignored for the tests you
would run; in the worst case, if you ran all the tests, all of the
default TestExpectations would be ignored. This was because
target was overloaded to also specify the "configuration" in
TestExpectations, and so you'd end up with a configuration that
was something other than "Release" or "Debug", but when the expectation
parser ran, it would only handle those two (an expectation with no
configuration specified would be treated as if both were, and still
wouldn't match).

This CL splits out the two concepts, so that the 'target' terminology
is used just to determine the last component of the build directory,
and the 'configuration' terminology is used to determine how to
apply TestExpectations.

However, the three sets of flags will still interact to determine
the behavior, as follows:

- if you specify none of --target, --release, or --debug, the
  default remains unchanged and works as-is (Release is used
  for both the target and the configuration).

- if you specify --target but not --release or --debug, then
  if the build directory is a GN build (as determined by looking
  for either args.gn or toolchain.ninja), we will default to a
  Debug configuration (the GN default), unless is_debug=false is
  in the args.gn file. Otherwise, if --target matches one of Debug,
  Debug_x64, Release, or Release_x64, we use the Debug or Release
  configuration as appropriate. If the target does not match
  one of the above strings and does not appear to point to a GN
  build, we error out with a (hopefully useful) message.

- if you specify --release or --debug but not --target, it works
  as before (using Release or Debug for both target and configuration
  as appropriate).

- if you specify both --target and either --release or --debug,
  and the target is a GN build for the wrong configuration, an
  exception is raised.

R=kbr@chromium.org
BUG=539507, 178207

Review URL: https://codereview.chromium.org/1720583002

Cr-Original-Commit-Position: refs/heads/master@{#378828}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: eccc853f65084a2f23d650ca38e009cfad0ab64f