Don't use a SubtreeLayoutScope object for SVG resource layout
SVG geometry depends on having its related resources laid out before
it can be laid out itself. This is enforced within the
layout code for the geometry, but doing so fires an assert in SubtreeLayoutScope
due to the fact that objects still need layout
when a child finishes. The assert is invalid, because the object is
about to be laid out so will not miss layout (we layout out of tree
order for resources).

This patch removes the SubtreeLayoutScope for resource layout, hence
avoiding the assert.

R=fs@opera.com
BUG=589389

Review URL: https://codereview.chromium.org/1749423005

Cr-Original-Commit-Position: refs/heads/master@{#378858}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c80d2b59c2357c6cae1dee9baef6fbaecb3c8136