PaymentRequest API.
An API that allows browsers to act as an intermediary between the three
key parties in a financial transaction: the merchant (e.g. an online web
store), the buyer (e.g. the user buying from the online web store), and
the Payment Method (e.g. credit card). Information necessary to process
and confirm a transaction is passed between the Payment Method and the
merchant via the browser with the buyer confirming and authorizing as
necessary across the flow.

Intent to implement:
https://groups.google.com/a/chromium.org/d/msg/blink-dev/gbSs15ZSWtA/JFU3H7fTDQAJ

BUG=587995

Review URL: https://codereview.chromium.org/1702223002

Cr-Original-Commit-Position: refs/heads/master@{#378911}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c757fcfb1cd39e90b7e520572ec8dcb276ebb748