Change implementation of createImageBitmap(Blob, premultiplyAlpha=false)
Currently, this function calls ImageDecoder to decode the blob with the
premultiplyAlpha to be always true, and then convert the underneath
SkImage to unpremultiplied format.

This CL makes change such that it calls ImageDecoder to decode the blob
with the premultiplyAlpha parameters, which will avoid the conversion
from premultiplied to unpremultiplied format.

BUG=589132

Review URL: https://codereview.chromium.org/1760683002

Cr-Original-Commit-Position: refs/heads/master@{#379022}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b443fe791b05a378bf2158239a69330bdff5a897