Partial implementation of subscription restrictions.
Currently, chrome requires that app developers provide a "sender_id" tag in the
manifest. This id is generated by the developer console and is not standard
for all browsers.

In the future, the app developer will be able to specify a public key for their
service, which will be registered with the push service and which the push
service can use to validate app services requesting to send messages.

BUG=583753

Review URL: https://codereview.chromium.org/1701313002

Cr-Original-Commit-Position: refs/heads/master@{#379045}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 71f5a08bb30153b56e8a3f9b447264a54b6d9c12