Mark usb/usbDevice.html as a slow LayoutTest.
I don't believe this test actually fails on bots. It simply does a
bunch of work and needs some extra time to run. Mark it as slow to see
if that makes the timeouts go away. If not then there is a real bug in
the test.

BUG=592183

Review URL: https://codereview.chromium.org/1775643002

Cr-Original-Commit-Position: refs/heads/master@{#379707}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ee57f6c40d33921df8364e4710a73620b57ccdb5