Avoid illegal definition of snprintf in VS 2015
VS 2015 supplies an implementation of snprintf which makes definining
it to _snprintf illegal - previously it was just a really bad idea.
This reinstates the compiler version check so that libxml will compile
in Chromium.

README.chromium was also updated to make it explicit that these checks
need to be added (for now) when config.h is copied.

A corresponding change will need to be made upstream.

BUG=440500,591920

Review URL: https://codereview.chromium.org/1770093002

Cr-Original-Commit-Position: refs/heads/master@{#379755}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f7bd300e8c52d6ebb7575a28d4cd9d3bdccaf439