WTF: Move TypeTraits.cpp to test-only target.
This patch renames TypeTraits.cpp to TypeTraitsTest.cpp and makes it
part of wtf_unittest instead of wtf target. Additionally, its content is
surrounded with anonymous namespace to indicate that the classes defined
there are not exposed in WTF namespace.

The file only contains static_assert assertions, so it doesn't have to
be compiled into wtf.

This change is a preparation of the fix for bug 592767.

BUG=592767

Review URL: https://codereview.chromium.org/1772163003

Cr-Original-Commit-Position: refs/heads/master@{#379800}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4a08542fb0f317809a3fca5ff092cfa60c1e52ac