Clear cached metadata when the resource data itself is invalidated.
Follow-up to crrev.com/1412413003: That change updated blink::Resource to
potentially update the Resource's data in-place when a revalidation
fails. That also invalidate that Resource's in-memory cached
meta data, which this change implements. This was previously
unnecessary, since a new Resource object would not have had that
cached data in the first place.

For the full drama, please refer to the referenced bug.

R=mkwst@chromium.org, jochen@chromium.org
BUG=chromium:589158

Review URL: https://codereview.chromium.org/1768143002

Cr-Original-Commit-Position: refs/heads/master@{#379828}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: cac81ec4db061ff13ccef6ca01d3522afeb0d814