Add layout tests for tex(Sub)Image2D for ImageBitmap
At this moment, there is an extensive coverage of this testing hosted on
KhronosGroup/WebGL. However, the problem is that we cannot have coverage
of the case for ImageBitmap constructor to take a "default" value for
some of the constructor options such as premultiplyAlpha. The reason is
that the option "default" is implementation-specific, and the results
could vary for different browsers.

In here, we add layout tests in chromium code base to have more test
cases than the ones hosted on KhronosGroup/WebGL.

Also, the KhronosGroup/WebGL testing suite tests tex(Sub)Image2D for 38
different kinds of pixel formats such as RGBA + UNSIGNED_SHORT_4444.
Here we test the basic one which is RGBA + UNSIGNED_BYTE, we can always
extend that in the future.

Review URL: https://codereview.chromium.org/1767113003

Cr-Original-Commit-Position: refs/heads/master@{#379879}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 7be4202c98d629b7eab64c251eb3ae3914e584e7