WTF: Fix Vector<T> memcpy'ing incorrectly for some T.
The core issue here is the implementation of IsTriviallyCopyAssignable
and IsTriviallyMoveAssignable was incorrect. They both use a compiler
builtin __has_trivial_assign(T), but it returns true if T's copy
assignment operator is declared as deleted.

To workaround that, a new type trait IsAssignable<T, From> is
implemented in the same way as std::is_assignable (as it's not available
at this time), and the assignability is used to limit the scope of
IsTrivially{Copy,Move}Assignable.

IsMoveAssignable isn't really right even with my patch for the reason
stated in the comments in the patch. I think it's impossible to
workaround that, though.

A bunch of static_asserts are added to test some invariants related to
this change.

BUG=592767

Review URL: https://codereview.chromium.org/1771303002

Cr-Original-Commit-Position: refs/heads/master@{#380329}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d7225e41c5557ef72e5ca439fcf8287be37e7c0f