Fix data race problem in createImageBitmap(Blob)
In createImageBitmap(Blob), the method creates an ImageDecoder
in a background thread, does the image decoding in the same thread, and
pass the ImageDecoder back to the main thread for creating an ImageBitmap.
This is causing the data race problem because the ImageDecoder
does some non-atomic reference counts internally.

This CL should fix the data race problem. In particular, we let the
ImageDecoder does image decoding in the background thread, and creates
a SkImage from the ImageDecoder, and pass the SkImage across threads.
Because SkImage is thread-safe RefCnted, this solves the data race
problem. In order to let postTask taking a SkImage, this CL changes
the CrossThreadCopier so that it can take an object that is SkRefCnt.

Locally I was able to repro the test case reported in the bug by
using the tsan build. With this CL the data race goes away.
Also, an existing layout tests that have data race condition
is changed.

BUG=593224

Review URL: https://codereview.chromium.org/1781613002

Cr-Original-Commit-Position: refs/heads/master@{#380609}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 283d18607e366095c377ddf72d196a8c4e256429