Fix SVGDocument return title that is not a child of the root element
The title element should be a child of the root element.
But currently return invalid title that is not a child of
the root element.

Because on getting, if the root element is an svg element in the SVG
namespace, then let value ba a concatenation of the data of all
the child Text nodes of the first title element in the SVG namespace
that is a child of the root element.[1]

And on setting, let element be the first title element in the SVG
namespace that is a child of the root element, if any.
If there isn't one, create a title element in the SVG namespace,
insert it as the first child of the root element, and let element
be that element.[1]

About a patch reverted[2][3], didn't consider inner svg document
in <html> or etc but is not <svg>.
If be inner svg document and insert <title> element that is not
svg namespace two times, occurred to crash.
Because call the member function of instance(m_titleElement)
that is null.
And added crash case that is
set-title-element-on-inner-svg-document.html.

[1] https://html.spec.whatwg.org/multipage/dom.html#document.title
[2] https://codereview.chromium.org/1771703003/
[3] https://codereview.chromium.org/1752713002/

BUG=543061, 592323

Review URL: https://codereview.chromium.org/1774913003

Cr-Original-Commit-Position: refs/heads/master@{#380634}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 962d172872d7097c7074c6597fd52732882a1d9f