Complete the implementation of up/down-mixing rules for AudioNode
The majority of up/down-mixing is not implemented in AudioBus.h/cpp.
This CL is to implement missing rules and  improves an existing layout
test for various channel configurations to challenge the implementation
extensively.

The spec for the mixing rule is here:
http://webaudio.github.io/web-audio-api/#channel-up-mixing-and-down-mixing

BUG=460958, 465908
TEST=LayoutTests/webaudio/audionode-channel-rules.html

Review URL: https://codereview.chromium.org/1773973002

Cr-Original-Commit-Position: refs/heads/master@{#380790}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c5bf1a9a84aadcf927f94e806f4647d65e3da2c1