Make pan-scrolling latch to a single scroller
It seems that having panScroll() use scrollByRecursively() was the instigator of the problem. In the case of pan scrolling and based on the expected behavior, once in the "pan-scrolling" mode, only the element in that mode should be scrolled. It should not recursively scroll the parent content.

When the current element scrolled to its farthest extent (in whatever direction) and there was a remaining scroll offset, scrollByRecursively would blindly seek out another scrollable container and try to scroll it. The problem would happen that the AutoscrollController::updateAutoscrollLayoutObject() method would, in some cases, fail to locate an outer scrollable container. This would then set m_autoscrollLayoutObject to nullptr, yet would remain in the AutoscrollForPanXXXX mode. Will need to dig further to better understand why.

Disclaimer: Since I'm new to the codebase, the following analysis is likely to contain many flaws merely due to my lack of familiarity with the code.

BUG=560148

Review URL: https://codereview.chromium.org/1781713002

Cr-Original-Commit-Position: refs/heads/master@{#380832}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: dca50fef52b59c63ac7a0d88bd37e59242cc1975